+++ extensions/riscv/rvr/c/rv64io_callbacks.h
@@ -7,8 +7,8 @@
  * consumers (HINT_STOREW, HINT_BUFFER) and public-values stores routed through
  * openvm_reveal. Backed by a dispatch table installed by
  * `Rv64IoExtension::register_host_callbacks`. */
-void openvm_hint_storew(uint32_t dest_addr);
-void openvm_hint_buffer(uint32_t dest_addr, uint32_t num_words);
-void openvm_reveal(uint64_t src_val, uint32_t ptr, uint32_t offset, uint32_t width);
+void openvm_hint_storew(uint64_t dest_addr);
+void openvm_hint_buffer(uint64_t dest_addr, uint32_t num_words);
+void openvm_reveal(uint64_t src_val, uint64_t ptr, uint32_t offset, uint32_t width);
 
 #endif /* RV64IO_CALLBACKS_H */
